Phase-II/backend/pyproject.toml
[project]
name = "hackathon-todo-backend"
version = "0.1.0"
description = "FastAPI backend for Phase II Todo Application"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.128.5",
    "sqlmodel==0.0.32",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "uvicorn[standard]>=0.30.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    "email-validator>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing"
